+++ lib/otr-activerecord.rb
@@ -1,7 +1,13 @@
 require 'active_record'
 require 'hashie-forbidden_attributes'
+
 require 'otr-activerecord/version'
 require 'otr-activerecord/activerecord'
-require 'otr-activerecord/middleware/connection_management'
-require 'otr-activerecord/middleware/query_cache'
-require 'otr-activerecord/defaults'
+
+ENV["DISABLE_DATABASE_ENVIRONMENT_CHECK"] ||= "true"
+
+OTR::ActiveRecord.db_dir = 'db'
+OTR::ActiveRecord.migrations_paths = %w(db/migrate)
+OTR::ActiveRecord.fixtures_path = 'test/fixtures'
+OTR::ActiveRecord.seed_file = 'seeds.rb'
+OTR::ActiveRecord.shim = OTR::ActiveRecord::Shim.new