+++ lib/otr-activerecord/middleware/query_cache.rb
@@ -5,28 +5,15 @@ module OTR
     #
     class QueryCache
       def initialize(app)
-        @handler = case ::ActiveRecord::VERSION::MAJOR
-                   when 4 then ::ActiveRecord::QueryCache.new(app)
-                   when 5, 6, 7 then ActionDispatchHandler.new(app)
-                   end
+        @app = app
       end
 
       def call(env)
-        @handler.call(env)
-      end
-
-      class ActionDispatchHandler
-        def initialize(app)
-          @app = app
-        end
-
-        def call(env)
-          state = nil
-          state = ::ActiveRecord::QueryCache.run
-          @app.call(env)
-        ensure
-          ::ActiveRecord::QueryCache.complete(state) if state
-        end
+        state = nil
+        state = ::ActiveRecord::QueryCache.run
+        @app.call(env)
+      ensure
+        ::ActiveRecord::QueryCache.complete(state) if state
       end
     end
   end