+++ spec/spec_helper.rb
@@ -1,7 +1,0 @@
-require 'otr-activerecord'
-
-RSpec.configure do |config|
-  config.color = true
-  config.tty = true
-  config.filter_run_when_matching :focus
-end