+++ test/support/active_record_models.rb
@@ -1,0 +1,7 @@
+class Category < ActiveRecord::Base
+  has_many :widgets
+end
+
+class Widget < ActiveRecord::Base
+  belongs_to :category
+end