+++ test/test_helper.rb
@@ -1,0 +1,5 @@
+require 'otr-activerecord'
+require 'rack'
+require 'minitest/autorun'
+
+Dir.glob('./test/support/*.rb').each { |file| require file }